cleanup(core): switch shell completion to yargs's .completion() API
Previously bin/nx.ts intercepted --get-yargs-completions and bypassed yargs.
Now yargs's .completion() callback handles the request — auto-enumerating
commands, subcommands, and options — while we layer project/target completions
on top via the 4-arg fallback callback.

- bin/nx.ts: early shortcut that skips workspace init but lets yargs handle
  the parse
- nx-commands.ts: parser-config carve-out (strip-dashed off for completion
  runs), declare the magic flag as boolean so yargs's parser doesn't eat the
  next arg, and register the completion callback
- completion/scripts.ts: new file with bash/zsh/fish script generators,
  lazy-loaded by command-object.ts via handleImport
- completion/routing.ts: project/target routing brain, renamed from
  fast-complete.ts (no longer the bin entry)
- completion/command-object.ts: slim to yargs wiring + handleImport
- completion-handler.ts: deleted (logic absorbed into routing.ts)
